fix: TeamRecommendation이 없을때 팀이 삭제되지 않는 문제 해결

+++ src/main/java/site/codemonster/comon/domain/recommendation/service/TeamRecommendationHighService.java
@@ -86,6 +86,9 @@ public class TeamRecommendationHighService {
     public void deleteTeamRecommendation(Long teamId) {
 
         Team team = teamLowService.findById(teamId);
+
+        if (!teamRecommendationLowService.isExistByTeam(team)) return;
+
         TeamRecommendation teamRecommendation = teamRecommendationLowService.findByTeam(team);
 
         platformRecommendationLowService.deleteByTeamRecommendationId(teamRecommendation.getId());
